fix: 졸업규정 편집 페이지 분리
