+++ app/[locale]/academics/undergraduate/degree-requirements/edit/page.tsx
@@ -1,0 +1,9 @@
+import { getDegreeRequirements } from '@/apis/academics';
+
+import DegreeRequirementsEditPageContent from './DegreeRequirementsEditPageContent';
+
+export default async function DegreeRequirementsEditPage() {
+  const data = await getDegreeRequirements();
+
+  return <DegreeRequirementsEditPageContent data={data} />;
+}
